+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - functions `not`, `and`, `or`, `xor`, `bool`, `inc`, `dec`, `crash`, `eq`, `eqAny`, `noop`
 - better (de)serialization: function/methods `serialize`, `deserialize`, `deserializeOrCrash` and `deserializeUnsafe`
 - implemented Jest snapshot serializer
+- `optArrayOpt` and `noneIfEmpty`
 
 ### Changed
 - improved return type of `actToOpt` method and function